Allow for extensibility in commonly-changing elements
As written, the DFXML schema allowed for no extensibility of elements,
including from new namespaces.  To allow a valid .dfxml file to include
non-"dfxml:" namespaced elements and attributes, the `<any>` element
must be included in the `<element>` definition.

The elements and attributes extended in this commit are the minimum
needed for differential annotations as idifference.py currently
generates.

Signed-off-by: Alex Nelson <a.nelson@prometheuscomputing.com>