Allow other-namespace attributes on all DFXML elements
Somebody asked if the all elements could be made extensible:
<https://github.com/dfxml-working-group/dfxml_schema/pull/9#issuecomment-25035429>

In response, this patch defines wrapper types for all the basic "xs:"
types DFXML elements use, adding an anyAttribute extension.  Now any
element in DFXML can have an attribute from other namespaces.

One extra change mixed in: partition_offset was a positiveInteger type.
It's legal for the offset to be 0, so this was changed to
nonNegativeInteger.